inspections: extract the `CheckDetails` component
